Commit semana 6 parte 2
Ultimos cambios semana 6. Whitelabel error mostrandose para la pagina de productos luego de los cambios, se va a ver en la proxima clase posible

+++ src/main/java/techtienda/controller/CategoriaController.java
@@ -41,7 +41,7 @@ public class CategoriaController {
     }
  
     @PostMapping("/eliminar")
-    public String eliminar(@RequestParam Integer idCategoria, RedirectAttributes redirectAttributes) {
+    public String eliminar(@RequestParam Long idCategoria, RedirectAttributes redirectAttributes) {
         String titulo="todoOk";
         String detalle="mensaje.eliminado";
         try {
@@ -61,7 +61,7 @@ public class CategoriaController {
     }
  
     @GetMapping("/modificar/{idCategoria}")    
-    public String modificar(@PathVariable("idCategoria") Integer idCategoria, Model model, RedirectAttributes redirectAttributes) {
+    public String modificar(@PathVariable("idCategoria") Long idCategoria, Model model, RedirectAttributes redirectAttributes) {
         Optional<Categoria> categoriaOpt = categoriaServices.getCategoria(idCategoria);
         if (categoriaOpt.isEmpty()) {
             redirectAttributes.addFlashAttribute("error", messageSource.getMessage("categoria.error01", null, Locale.getDefault()));
