+++ src/main/java/techtienda/services/CategoriaServices.java
@@ -24,7 +24,7 @@ public class CategoriaServices {
         return categoriaRepository.findAll();
     }
     @Transactional(readOnly = true)
-    public Optional<Categoria> getCategoria(Integer idCategoria) {
+    public Optional<Categoria> getCategoria(Long idCategoria) {
         return categoriaRepository.findById(idCategoria);
     }
  
@@ -48,7 +48,7 @@ public class CategoriaServices {
     }
  
     @Transactional
-    public void delete(Integer idCategoria) {
+    public void delete(Long idCategoria) {
         // Verifica si la categoría existe antes de intentar eliminarlo
         if (!categoriaRepository.existsById(idCategoria)) {
             // Lanza una excepción para indicar que el usuario no fue encontrado
